[Site Isolation] Web Inspector: Make InspectorStyleSheet-related actions accessible for frame targets https://bugs.webkit.org/show_bug.cgi?id=316876
Reviewed by BJ Burg.

Move StyleSheetAction and subclasses out of InspectorCSSAgent's
file-private nested scope into InspectorStyleSheet.h, so
FrameCSSAgent can construct and submit the same actions to its
InspectorHistory when we later implement style-editing commands on
frame targets.

These classes only depend on InspectorStyleSheet and InspectorHistory;
there is no agent-specific state, so the original nesting inside
InspectorCSSAgent was incidental.

Also address safer-cpp issues by holding RefPtr on the stack whenever
appropriate.

No new tests; no change in observable behavior.

* Source/WebCore/inspector/InspectorStyleSheet.h:
(WebCore::StyleSheetAction::StyleSheetAction):
(isType):
* Source/WebCore/inspector/agents/InspectorCSSAgent.cpp:
(WebCore::InspectorCSSAgent::StyleSheetAction::StyleSheetAction): Deleted.
(): Deleted.
(isType): Deleted.
* Source/WebCore/inspector/agents/InspectorCSSAgent.h:

Canonical link: https://commits.webkit.org/315764@main
