+++ Source/WebCore/inspector/agents/InspectorCSSAgent.cpp
@@ -91,200 +91,6 @@ using namespace Inspector;
 
 WTF_MAKE_TZONE_ALLOCATED_IMPL(InspectorCSSAgent);
 
-class InspectorCSSAgent::StyleSheetAction : public InspectorHistory::Action {
-    WTF_MAKE_NONCOPYABLE(StyleSheetAction);
-public:
-    explicit StyleSheetAction(InspectorStyleSheet* styleSheet)
-        : InspectorHistory::Action()
-        , m_styleSheet(styleSheet)
-    {
-    }
-
-protected:
-    RefPtr<InspectorStyleSheet> m_styleSheet;
-};
-
-class InspectorCSSAgent::SetStyleSheetTextAction final : public InspectorCSSAgent::StyleSheetAction {
-    WTF_MAKE_NONCOPYABLE(SetStyleSheetTextAction);
-public:
-    SetStyleSheetTextAction(InspectorStyleSheet* styleSheet, const String& text)
-        : InspectorCSSAgent::StyleSheetAction(styleSheet)
-        , m_text(text)
-    {
-    }
-
-private:
-    bool isSetStyleSheetTextAction() const final { return true; }
-
-    ExceptionOr<void> perform() final
-    {
-        auto result = protect(m_styleSheet)->text();
-        if (result.hasException())
-            return result.releaseException();
-        m_oldText = result.releaseReturnValue();
-        return redo();
-    }
-
-    ExceptionOr<void> undo() final
-    {
-        auto result = protect(m_styleSheet)->setText(m_oldText);
-        if (result.hasException())
-            return result.releaseException();
-        protect(m_styleSheet)->reparseStyleSheet(m_oldText);
-        return { };
-    }
-
-    ExceptionOr<void> redo() final
-    {
-        auto result = protect(m_styleSheet)->setText(m_text);
-        if (result.hasException())
-            return result.releaseException();
-        protect(m_styleSheet)->reparseStyleSheet(m_text);
-        return { };
-    }
-
-    String mergeId() final
-    {
-        return makeString("SetStyleSheetText "_s, m_styleSheet->id());
-    }
-
-    void merge(std::unique_ptr<Action> action) override
-    {
-        ASSERT(action->mergeId() == mergeId());
-        m_text = downcast<SetStyleSheetTextAction>(*action).m_text;
-    }
-
-    String m_text;
-    String m_oldText;
-};
-
-class InspectorCSSAgent::SetStyleTextAction final : public InspectorCSSAgent::StyleSheetAction {
-    WTF_MAKE_NONCOPYABLE(SetStyleTextAction);
-public:
-    SetStyleTextAction(InspectorStyleSheet* styleSheet, const InspectorCSSId& cssId, const String& text)
-        : InspectorCSSAgent::StyleSheetAction(styleSheet)
-        , m_cssId(cssId)
-        , m_newStyleDeclarationText(text)
-    {
-    }
-
-    ExceptionOr<void> perform() override
-    {
-        return redo();
-    }
-
-    ExceptionOr<void> undo() override
-    {
-        return protect(m_styleSheet)->setRuleStyleText(
-            m_cssId,
-            m_oldStyleDeclarationText,
-            nullptr, /* outOldStyleDeclarationText */
-            &m_oldRuleText,
-            nullptr /* outOldRuleText */);
-    }
-
-    ExceptionOr<void> redo() override
-    {
-        return protect(m_styleSheet)->setRuleStyleText(
-            m_cssId,
-            m_newStyleDeclarationText,
-            &m_oldStyleDeclarationText,
-            nullptr, /* newRuleText */
-            &m_oldRuleText);
-    }
-
-    String mergeId() override
-    {
-        ASSERT(m_styleSheet->id() == m_cssId.styleSheetId());
-        return makeString("SetStyleText "_s, m_styleSheet->id(), ':', m_cssId.ordinal());
-    }
-
-    void merge(std::unique_ptr<Action> action) override
-    {
-        ASSERT(action->mergeId() == mergeId());
-
-        SetStyleTextAction* other = static_cast<SetStyleTextAction*>(action.get());
-        m_newStyleDeclarationText = other->m_newStyleDeclarationText;
-    }
-
-private:
-    InspectorCSSId m_cssId;
-    String m_newStyleDeclarationText;
-    String m_oldStyleDeclarationText;
-    String m_oldRuleText;
-};
-
-class InspectorCSSAgent::SetRuleHeaderTextAction final : public InspectorCSSAgent::StyleSheetAction {
-    WTF_MAKE_NONCOPYABLE(SetRuleHeaderTextAction);
-public:
-    SetRuleHeaderTextAction(InspectorStyleSheet* styleSheet, const InspectorCSSId& cssId, const String& newHeaderText)
-        : InspectorCSSAgent::StyleSheetAction(styleSheet)
-        , m_cssId(cssId)
-        , m_newHeaderText(newHeaderText)
-    {
-    }
-
-private:
-    ExceptionOr<void> perform() final
-    {
-        auto result = protect(m_styleSheet)->ruleHeaderText(m_cssId);
-        if (result.hasException())
-            return result.releaseException();
-        m_oldHeaderText = result.releaseReturnValue();
-        return redo();
-    }
-
-    ExceptionOr<void> undo() final
-    {
-        return protect(m_styleSheet)->setRuleHeaderText(m_cssId, m_oldHeaderText);
-    }
-
-    ExceptionOr<void> redo() final
-    {
-        return protect(m_styleSheet)->setRuleHeaderText(m_cssId, m_newHeaderText);
-    }
-
-    InspectorCSSId m_cssId;
-    String m_newHeaderText;
-    String m_oldHeaderText;
-};
-
-class InspectorCSSAgent::AddRuleAction final : public InspectorCSSAgent::StyleSheetAction {
-    WTF_MAKE_NONCOPYABLE(AddRuleAction);
-public:
-    AddRuleAction(InspectorStyleSheet* styleSheet, const String& selector)
-        : InspectorCSSAgent::StyleSheetAction(styleSheet)
-        , m_selector(selector)
-    {
-    }
-
-    InspectorCSSId NODELETE newRuleId() const { return m_newId; }
-
-private:
-    ExceptionOr<void> perform() final
-    {
-        return redo();
-    }
-
-    ExceptionOr<void> undo() final
-    {
-        return protect(m_styleSheet)->deleteRule(m_newId);
-    }
-
-    ExceptionOr<void> redo() final
-    {
-        auto result = protect(m_styleSheet)->addRule(m_selector);
-        if (result.hasException())
-            return result.releaseException();
-        m_newId = protect(m_styleSheet)->ruleOrStyleId(result.releaseReturnValue());
-        return { };
-    }
-
-    InspectorCSSId m_newId;
-    String m_selector;
-    String m_oldSelector;
-};
-
 InspectorCSSAgent::InspectorCSSAgent(PageAgentContext& context)
     : InspectorAgentBase("CSS"_s, context)
     , m_frontendDispatcher(makeUniqueRef<CSSFrontendDispatcher>(context.frontendRouter))
@@ -1471,10 +1277,3 @@ void InspectorCSSAgent::resetPseudoStates()
 }
 
 } // namespace WebCore
-
-SPECIALIZE_TYPE_TRAITS_BEGIN(WebCore::InspectorCSSAgent::SetStyleSheetTextAction)
-    static bool isType(const WebCore::InspectorHistory::Action& action)
-    {
-        return action.isSetStyleSheetTextAction();
-    }
-SPECIALIZE_TYPE_TRAITS_END()
